requirements.txt
pytube
whisper.ai